Renamed package from org.orienteer to org.orienteer.mail
