+++ orienteer-mail/src/main/java/org/orienteer/mail/Initializer.java
@@ -1,4 +1,4 @@
-package org.orienteer;
+package org.orienteer.mail;
 
 import org.apache.wicket.Application;
 import org.apache.wicket.IInitializer;
@@ -18,6 +18,7 @@ public class Initializer implements IInitializer
 	@Override
 	public void destroy(Application application) {
 		OrienteerWebApplication app = (OrienteerWebApplication)application;
+		app.unregisterModule(OMailModule.class);
 	}
 	
 }
