+++ orienteer-mail/src/main/java/org/orienteer/mail/model/OMail.java
@@ -1,4 +1,4 @@
-package org.orienteer.model;
+package org.orienteer.mail.model;
 
 import com.orientechnologies.orient.core.db.record.OIdentifiable;
 import com.orientechnologies.orient.core.record.impl.ODocument;
