+++ orienteer-mail/src/main/java/org/orienteer/mail/model/OMailSettings.java
@@ -1,4 +1,4 @@
-package org.orienteer.model;
+package org.orienteer.mail.model;
 
 import com.orientechnologies.orient.core.record.impl.ODocument;
 import com.orientechnologies.orient.core.type.ODocumentWrapper;
