+++ orienteer-mail/src/main/java/org/orienteer/mail/package-info.java
@@ -1,4 +1,4 @@
 /**
  * Main package for 'orienteer-mail' module
  */
-package org.orienteer;
+package org.orienteer.mail;
