+++ orienteer-mail/src/main/java/org/orienteer/mail/service/package-info.java
@@ -1,4 +1,4 @@
 /**
  * Contains services for 'orienteer-mail' module
  */
-package org.orienteer.service;
+package org.orienteer.mail.service;
