+++ orienteer-mail/src/main/java/org/orienteer/mail/task/package-info.java
@@ -1,4 +1,4 @@
 /**
  * Contains task classes for 'orienteer-mail' module
  */
-package org.orienteer.task;
+package org.orienteer.mail.task;
