+++ orienteer-mail/src/main/java/org/orienteer/mail/util/package-info.java
@@ -1,4 +1,4 @@
 /**
  * Contains utility classes for 'orienteer-mail' module
  */
-package org.orienteer.util;
+package org.orienteer.mail.util;
