+++ orienteer-mail/src/main/resources/META-INF/services/org.apache.wicket.IInitializer
@@ -1,1 +1,1 @@
-org.orienteer.Initializer
+org.orienteer.mail.Initializer
