+++ orienteer-mail/src/test/java/org/orienteer/mail/TestSendMail.java
@@ -1,17 +1,18 @@
-package org.orienteer;
+package org.orienteer.mail;
 
 import com.orientechnologies.orient.core.sql.OCommandSQL;
 import org.junit.After;
+import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.orienteer.core.OrienteerWebApplication;
 import org.orienteer.junit.OrienteerTestRunner;
-import org.orienteer.model.OMail;
-import org.orienteer.model.OMailSettings;
-import org.orienteer.model.OPreparedMail;
-import org.orienteer.service.IOMailService;
-import org.orienteer.service.OMailServiceTest;
+import org.orienteer.mail.model.OMail;
+import org.orienteer.mail.model.OMailSettings;
+import org.orienteer.mail.model.OPreparedMail;
+import org.orienteer.mail.service.OMailServiceTest;
+import org.orienteer.mail.service.IOMailService;
 import ru.ydn.wicket.wicketorientdb.utils.DBClosure;
 
 import javax.mail.MessagingException;
@@ -20,8 +21,6 @@ import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.UUID;
-
-import static org.junit.Assert.assertEquals;
 
 @RunWith(OrienteerTestRunner.class)
 public class TestSendMail {
@@ -57,7 +56,7 @@ public class TestSendMail {
 
     @Test
     public void testSendSingleMail() throws UnsupportedEncodingException, MessagingException {
-        testService.onSendMail(mail -> assertEquals(preparedMail, mail))
+        testService.onSendMail(mail -> Assert.assertEquals(preparedMail, mail))
                 .sendMail(preparedMail);
     }
