+++ orienteer-mail/src/test/java/org/orienteer/mail/service/TestInitModule.java
@@ -1,4 +1,4 @@
-package org.orienteer.service;
+package org.orienteer.mail.service;
 
 import com.google.inject.AbstractModule;
