+++ orienteer-mail/src/test/resources/META-INF/services/com.google.inject.Module
@@ -1,1 +1,1 @@
-org.orienteer.service.TestInitModule
+org.orienteer.mail.service.TestInitModule
